Run KafkaSource smokes in upgrade test
